[INTERNAL] replaceGlobals Task finder improvement
Check that each found replacement is evaluated if replacement would
take place, e.g. jQuery calls which get replaced with itself.
These "replacements" are present (which do not do a thing) to be able
to manage the dependencies. Because these dependencies will be left as
is if there are findings, otherwise this dependency will be
removed/added if not present.